#154: feat 上下文组织接口与策略实现
