+++ docs/review/context-organizer-review.md
@@ -1,0 +1,194 @@
+# Context Organizer 增强 Review
+
+✌Bazinga！
+
+## 一、架构设计评估
+
+### 1.1 整体架构
+
+```
+┌─────────────────────────────────────────────────────────────┐
+│                    ContextManager                            │
+│                   (门面/协调者)                               │
+└──────────────────────────┬──────────────────────────────────┘
+                           │
+                           ▼
+┌─────────────────────────────────────────────────────────────┐
+│               ContextOrganizerFactory                        │
+│                   (工厂模式)                                 │
+└──────────────────────────┬──────────────────────────────────┘
+                           │
+           ┌───────────────┼───────────────┐
+           ▼               ▼               ▼
+    ┌──────────────┐ ┌──────────────┐ ┌──────────────┐
+    │FullOrganizer │ │SimpleOrganizer│ │  Future...   │
+    │  (完整策略)  │ │  (简单策略)   │ │   (可扩展)   │
+    └──────────────┘ └──────────────┘ └──────────────┘
+```
+
+**评价**: 采用策略模式+工厂模式，设计清晰，符合开闭原则（对扩展开放，对修改关闭）。
+
+### 1.2 接口设计
+
+```javascript
+// IContextOrganizer 接口
+- organize(memorySystem, userId, options): Promise<ContextResult>
+- getName(): string
+- getDescription(): string
+```
+
+**优点**:
+- 接口简洁，职责单一
+- `ContextResult` 统一返回结构，便于上层处理
+
+**潜在问题**:
+- 缺少配置验证方法
+- 缺少策略能力声明（如支持哪些特性）
+
+---
+
+## 二、策略实现评估
+
+### 2.1 FullContextOrganizer（完整上下文策略）
+
+| 特性 | 实现 | 评价 |
+|------|------|------|
+| 未归档消息 | `getUnarchivedMessages(userId, null)` | ✅ 不限制数量，依赖压缩机制 |
+| Inner Voices | 最近 3 条 | ✅ 合理数量 |
+| Topic 总结 | 最近 10 个，只取 active | ✅ 状态过滤正确 |
+| 用户档案 | 包含完整上下文 | ✅ |
+| Token 估算 | `estimateTokens()` | ✅ 便于监控 |
+
+**适用场景**: 需要完整上下文的深度对话、复杂任务处理
+
+### 2.2 SimpleContextOrganizer（简单上下文策略）
+
+| 特性 | 实现 | 评价 |
+|------|------|------|
+| 最近消息 | `getRecentMessages(userId, 10)` | ⚠️ 注意：这是所有消息，不是未归档 |
+| Inner Voices | 默认不包含 | ✅ 减少 token 消耗 |
+| Topic 总结 | 最近 5 个 | ✅ 轻量级 |
+| 用户档案 | 包含完整上下文 | ✅ |
+
+**适用场景**: 快速问答、轻量级对话
+
+**潜在问题**:
+- `getRecentMessages` 获取的是最近 N 条消息（包含已归档的），而 `FullContextOrganizer` 使用的是 `getUnarchivedMessages`
+- 这可能导致两种策略的行为不一致：Simple 模式下可能包含已归档消息
+
+---
+
+## 三、状态管理评估
+
+### 3.1 Topic 状态过滤
+
+**改进点**:
+- `getTopics(userId, limit, 'active')` - 正确过滤 active 状态
+- 防止返回已删除/归档的 topic
+
+**遗漏点**:
+- `createTopic` 时没有显式设置 `status: 'active'`
+- `compressContext` 创建 topic 时没有设置 `startTime`/`endTime`
+
+### 3.2 状态转换生命周期
+
+```
+         ┌──────────┐
+         │  active  │ ←── 新创建的 topic
+         └────┬─────┘
+              │
+              ▼
+         ┌──────────┐
+         │ archived │ ←── 被归档（当前未使用）
+         └────┬─────┘
+              │
+              ▼
+         ┌──────────┐
+         │ deleted  │ ←── 软删除（当前未使用）
+         └──────────┘
+```
+
+**建议**: 补充状态转换的完整逻辑和文档
+
+---
+
+## 四、Agent 系统适用性评估
+
+### 4.1 对 Agent 系统的积极影响
+
+| 方面 | 评价 |
+|------|------|
+| **上下文灵活性** | ✅ 不同场景可选择不同策略 |
+| **Token 优化** | ✅ Simple 模式显著减少 token 消耗 |
+| **可扩展性** | ✅ 新增策略无需修改现有代码 |
+| **测试性** | ✅ 策略独立，便于单元测试 |
+
+### 4.2 潜在风险
+
+| 风险 | 严重程度 | 建议 |
+|------|----------|------|
+| 重复代码 | 中 | ContextManager 和 BaseContextOrganizer 存在大量重复 |
+| 行为不一致 | 中 | Simple 模式使用 getRecentMessages 而非 getUnarchivedMessages |
+| 缺少切换机制 | 低 | 未实现策略动态切换的入口 |
+
+### 4.3 架构优化建议
+
+1. **消除重复代码**
+   - ContextManager 应该内部使用 ContextOrganizer，而非独立实现
+   - 或者：将 ContextManager 标记为 deprecated，统一使用 Organizer
+
+2. **统一消息获取逻辑**
+   ```javascript
+   // SimpleContextOrganizer 应该改为：
+   const recentMessages = await memorySystem.getUnarchivedMessages(userId, this.messageCount);
+   ```
+
+3. **添加策略切换入口**
+   ```javascript
+   // 在专家配置中添加
+   expert.context_strategy = 'full' | 'simple' | 'custom'
+   ```
+
+---
+
+## 五、测试建议
+
+### 5.1 单元测试覆盖
+
+- [ ] IContextOrganizer 接口契约测试
+- [ ] FullContextOrganizer 完整流程测试
+- [ ] SimpleContextOrganizer 简化流程测试
+- [ ] ContextOrganizerFactory 注册/获取测试
+- [ ] Topic 状态过滤测试
+
+### 5.2 集成测试
+
+- [ ] 两种策略的实际 token 消耗对比
+- [ ] 长对话场景下的行为差异
+- [ ] 状态过滤的正确性验证
+
+---
+
+## 六、总结
+
+### 优点
+1. 架构设计合理，采用策略模式便于扩展
+2. 接口定义清晰，返回结构统一
+3. 状态过滤实现正确
+4. 支持不同场景的上下文需求
+
+### 需要改进
+1. **消除重复代码** - ContextManager 和 Organizer 存在大量重复
+2. **统一消息获取逻辑** - Simple 模式应使用 getUnarchivedMessages
+3. **补充 Topic 创建逻辑** - 显式设置 status 和时间边界
+4. **完善文档** - 状态转换生命周期、策略选择指南
+
+### 建议 Issue #154 包含
+1. 重构 ContextManager 使用 Organizer（或标记 deprecated）
+2. 修复 SimpleOrganizer 的消息获取逻辑
+3. 补充 Topic 创建时的状态和时间设置
+4. 添加策略选择的配置入口
+
+---
+
+✌Bazinga！亲爱的
