+++ docs/review/full-context-organizer-analysis.md
@@ -1,0 +1,173 @@
+# FullContextOrganizer 深度分析报告
+
+## 一、系统背景
+
+Touwaka Mate 是一个 AI 专家副本系统，具有以下核心特性：
+- **Expert（专家）**：具有独特人设的 AI 角色
+- **Topic（话题）**：对话历史的阶段性总结
+- **Skill（技能）**：专家可调用的工具能力
+- **双心智架构**：表达心智 + 反思心智（Inner Voice）
+
+## 二、压缩机制分析（关键流程）
+
+### 2.1 压缩触发时机
+
+压缩在每次用户发消息后、构建上下文之前自动执行：
+
+```javascript
+// chat-service.js:169-190
+// 6. 检查是否需要压缩上下文（在 buildContext 之前！）
+const compressionCheck = await expertService.memorySystem.shouldCompressContext(
+  user_id,
+  expertService.getDefaultModelConfig().max_tokens || 128000,
+  expertService.expertConfig?.expert?.context_threshold || 0.7,
+  5,   // 最小消息数：5条消息后开始检查压缩
+  50   // 最大未归档消息数：超过50条强制压缩
+);
+
+if (compressionCheck.needCompress) {
+  // 同步执行压缩（阻塞，但必要）
+  const compressResult = await expertService.memorySystem.compressContext(user_id, {...});
+}
+
+// 7. 构建上下文（压缩后执行）
+const context = await expertService.buildContext(user_id, content, topic_id, taskContext);
+```
+
+### 2.2 压缩触发条件
+
+```javascript
+// memory-system.js:277-322
+async shouldCompressContext(userId, contextSize = 128000, threshold = 0.7, minMessages = 5, maxMessages = 50) {
+  // 条件1：Token 超阈值
+  if (estimatedTokens >= tokenThreshold) {
+    return { needCompress: true, reason: `Token 数 ${estimatedTokens} >= 阈值 ${tokenThreshold}` };
+  }
+
+  // 条件2：消息数超限（硬限制，防止 Token 计算不准）
+  if (unarchivedMessages.length >= maxMessages) {
+    return { needCompress: true, reason: `未归档消息数 ${unarchivedMessages.length} >= 最大值 ${maxMessages}` };
+  }
+}
+```
+
+### 2.3 反思触发的强制压缩
+
+```javascript
+// chat-service.js:1278-1294
+// 当反思检测到话题偏移时，强制压缩
+if (reflection.topicSuggestion?.shouldCreateNew) {
+  const compressResult = await this.memorySystem.compressContext(user_id, {
+    force: true,  // 强制压缩，跳过阈值检查
+  });
+}
+```
+
+### 2.4 结论
+
+**压缩机制是可靠的：**
+- ✅ 在 `buildContext` 之前执行
+- ✅ 有双重保护：Token 阈值 + 消息数硬限制（50条）
+- ✅ 反思可触发强制压缩
+- ✅ 不会出现大量未归档消息的情况
+
+**FullContextOrganizer 获取的"全部未归档消息"实际上最多只有 50 条。**
+
+## 三、当前 FullContextOrganizer 实现分析
+
+### 3.1 核心流程
+
+```
+1. 获取全部未归档消息（实际最多 50 条）
+2. 获取 Inner Voices（默认 3 条）
+3. 获取 Topic 总结（默认 10 个）
+4. 构建系统提示（Soul + 技能 + 任务上下文 + RAG + Inner Voices）
+5. 返回 ContextResult
+```
+
+### 3.2 上下文组成结构
+
+| 组件 | 来源 | Token 占用（估算） | 优先级 |
+|------|------|-------------------|--------|
+| System Prompt | expert.prompt_template | 200-500 | 固定 |
+| Timestamp | 系统生成 | ~100 | 固定 |
+| Soul | expert 核心属性 | 300-800 | 固定 |
+| Skills | 技能列表 | 200-1000 | 固定 |
+| Task Context | 任务工作空间 | 300-500 | 可选 |
+| Topic Summaries | topics 表 | 500-2000 | 可调整 |
+| RAG Context | 知识库检索 | 0-3000 | 可选 |
+| Inner Voices | messages.inner_voice | 300-600 | 可调整 |
+| User Info Guidance | user_profiles | 0-100 | 可选 |
+| History Messages | messages 表 | **不固定（最多50条）** | 核心 |
+
+## 四、潜在的优化空间
+
+### 4.1 🟡 P2 - Topic 数量固定
+
+**现状**：固定获取 10 个 topic
+
+**潜在问题**：
+- 长期用户可能有几十个 topic，只能看到最近 10 个
+- 旧 topic 可能不再相关但仍占用 token
+
+**优化建议**：
+- 根据可用 token 预算动态调整 topic 数量
+- 或者考虑 topic 相关性排序
+
+### 4.2 🟡 P2 - 缺少 Token 预算分配
+
+**现状**：各组件没有明确的 token 预算分配
+
+**潜在问题**：
+- 当消息接近 50 条时，加上 10 个 topic + 3 个 inner voice + RAG，可能接近上下文窗口
+- 没有优先级机制
+
+**优化建议**：
+- 为各组件设置 token 预算上限
+- 实现动态调整机制
+
+### 4.3 🟢 P3 - 性能优化
+
+**现状**：每次 `buildContext` 都重新构建完整的系统提示
+
+**优化建议**：
+- 缓存不变的 Soul、Skills 等内容
+- 仅更新动态部分（Inner Voices、Topic Summaries）
+
+## 五、总结
+
+### 压缩机制评估
+
+**结论：压缩机制设计合理，能有效控制未归档消息数量。**
+
+| 机制 | 状态 | 说明 |
+|------|------|------|
+| Token 阈值触发 | ✅ 正常 | 默认 70%，可配置 |
+| 消息数硬限制 | ✅ 正常 | 默认 50 条，强制触发 |
+| 反思强制压缩 | ✅ 正常 | 话题偏移时触发 |
+| 执行时机 | ✅ 正确 | buildContext 之前 |
+
+### FullContextOrganizer 评估
+
+**结论：当前实现满足系统需求，有少量优化空间但不紧急。**
+
+| 方面 | 评估 | 说明 |
+|------|------|------|
+| 功能完整性 | ✅ | 正确获取消息、Inner Voices、Topic Summaries |
+| Token 控制 | ✅ | 依赖压缩机制，最多 50 条消息 |
+| 性能 | ⚠️ | 可优化，但不是瓶颈 |
+| 灵活性 | ⚠️ | 参数固定，可考虑动态调整 |
+
+### 建议
+
+1. **当前无需紧急修改** - 压缩机制工作正常
+2. **可选优化**：
+   - 动态 Topic 数量（根据预算调整）
+   - 缓存不变的系统提示内容
+3. **SimpleContextOrganizer** 作为轻量级替代方案，适合不需要完整上下文的场景
+
+---
+
+*分析时间：2026-03-15*
+*分析者：Maria*
+*修订：根据实际压缩机制流程修正分析*
