+++ lib/chat-service.js
@@ -627,6 +627,11 @@ class ChatService {
   /**
    * 保存工具消息
    * 每个工具调用完成后立即保存，实现增量持久化
+   *
+   * 设计原则：
+   * - 存储层完整保存所有信息，不做截断
+   * - 上下文构建时根据场景决定使用多少内容
+   *
    * @param {string} topic_id - 话题ID（用于关联，但消息 topic_id 为 NULL）
    * @param {string} user_id - 用户ID
    * @param {object} toolResult - 工具执行结果
@@ -636,22 +641,20 @@ class ChatService {
   async saveToolMessage(topic_id, user_id, toolResult, expert_id = null) {
     const message_id = Utils.newID(20);
 
-    // 构建 tool_calls 字段内容（复用现有字段）
+    // 构建 tool_calls 字段内容（完整存储）
     // 格式: { tool_call_id, name, arguments, success, duration, timestamp, context }
-    // 注意：不存储 content（工具返回结果），因为可能很大
     const toolCallsData = {
       tool_call_id: toolResult.toolCallId,
       name: toolResult.toolName,
-      arguments: toolResult.arguments || null,  // 工具调用参数
+      arguments: toolResult.arguments || null,  // 工具调用参数（完整）
       success: toolResult.success,
       duration: toolResult.duration || 0,
       timestamp: new Date().toISOString(),
       context: toolResult.context || null,  // 工具调用前的状态文本上下文
     };
 
-    // 对 content 进行截断（工具返回结果可能很大）
-    // 截断后存储到 content 字段，用于前端展示和 LLM 上下文
-    const MAX_CONTENT_LENGTH = 1000;  // 最大 1000 字符
+    // 完整存储工具返回结果（不再截断）
+    // 上下文构建时由 ContextOrganizer 决定使用多少内容
     let content = '';
     if (toolResult.data !== undefined) {
       content = typeof toolResult.data === 'string'
@@ -660,12 +663,6 @@ class ChatService {
     } else if (toolResult.error) {
       content = toolResult.error;
     }
-    
-    // 截断过长的内容
-    if (content.length > MAX_CONTENT_LENGTH) {
-      content = content.substring(0, MAX_CONTENT_LENGTH) +
-        `\n...[已截断，原始 ${content.length} 字符]`;
-    }
 
     await this.Message.create({
       id: message_id,
@@ -673,11 +670,11 @@ class ChatService {
       user_id,
       expert_id,
       role: 'tool',
-      content,  // 截断后的工具返回内容
-      tool_calls: JSON.stringify(toolCallsData),  // 工具元数据（含 arguments）
+      content,  // 完整的工具返回内容（不再截断）
+      tool_calls: JSON.stringify(toolCallsData),  // 工具元数据（含完整 arguments）
     });
 
-    logger.debug(`[ChatService] 工具消息已保存: ${toolResult.toolName}, message_id: ${message_id}`);
+    logger.debug(`[ChatService] 工具消息已保存: ${toolResult.toolName}, message_id: ${message_id}, content_length: ${content.length}`);
     return message_id;
   }
