+++ lib/db.js
@@ -572,7 +572,7 @@ class Database {
   async createTopic(topicData) {
     const {
       id, expertId, userId, name, description,
-      category, startTime, endTime
+      category, startTime, endTime, status
     } = topicData;
 
     return await this.models.topic.create({
@@ -582,18 +582,32 @@ class Database {
       title: name,
       description,
       category,
+      status: status || 'active',  // 默认 active
+      start_time: startTime || new Date(),
+      end_time: endTime || new Date(),
     });
   }
 
   /**
    * 获取用户的 Topics（支持不同签名）
+   * @param {string} expertId - 专家ID
+   * @param {string} userId - 用户ID
+   * @param {number} limit - 数量限制
+   * @param {string} status - 状态过滤（默认 'active'，传 null 表示不过滤）
    */
-  async getTopicsByExpertAndUser(expertId, userId, limit = 10) {
+  async getTopicsByExpertAndUser(expertId, userId, limit = 10, status = 'active') {
+    const where = {
+      expert_id: expertId,
+      user_id: userId,
+    };
+
+    // 如果指定了状态，添加状态过滤
+    if (status) {
+      where.status = status;
+    }
+
     return await this.models.topic.findAll({
-      where: {
-        expert_id: expertId,
-        user_id: userId,
-      },
+      where,
       attributes: ['id', 'expert_id', 'user_id', 'title', 'description', 'category', 'status', 'message_count', 'created_at', 'updated_at'],
       order: [['updated_at', 'DESC']],
       limit,
