+++ lib/memory-system.js
@@ -203,10 +203,11 @@ class MemorySystem {
    * 获取用户的 Topics
    * @param {string} userId - 用户ID
    * @param {number} limit - 数量限制
+   * @param {string} status - 状态过滤（默认 'active'，传 null 表示不过滤）
    * @returns {Promise<Array>} Topic 列表
    */
-  async getTopics(userId, limit = 10) {
-    return await this.db.getTopicsByExpertAndUser(this.expertId, userId, limit);
+  async getTopics(userId, limit = 10, status = 'active') {
+    return await this.db.getTopicsByExpertAndUser(this.expertId, userId, limit, status);
   }
 
   /**
@@ -417,7 +418,19 @@ class MemorySystem {
       for (const topic of topics) {
         const topicId = this.generateTopicId();
 
-        // 创建 Topic
+        // 获取该话题的消息
+        const topicMessages = unarchivedMessages.slice(topic.startIndex, topic.endIndex + 1);
+        const messageIds = topicMessages.map(m => m.id);
+
+        // 计算话题的时间范围
+        const startTime = topicMessages.length > 0 
+          ? new Date(topicMessages[0].timestamp) 
+          : new Date();
+        const endTime = topicMessages.length > 0 
+          ? new Date(topicMessages[topicMessages.length - 1].timestamp) 
+          : new Date();
+
+        // 创建 Topic（显式设置 status 和时间边界）
         await this.db.createTopic({
           id: topicId,
           expertId: this.expertId,
@@ -425,12 +438,12 @@ class MemorySystem {
           name: topic.title,
           description: topic.summary,
           category: topic.category || 'general',
+          status: 'active',  // 显式设置状态
+          startTime,
+          endTime,
         });
 
-        // 获取该话题的消息 ID
-        const messageIds = unarchivedMessages
-          .slice(topic.startIndex, topic.endIndex + 1)
-          .map(m => m.id);
+        logger.debug(`[MemorySystem] 创建话题: ${topic.title}, 时间范围: ${startTime.toISOString()} ~ ${endTime.toISOString()}`);
 
         // 关联消息到 Topic
         if (messageIds.length > 0) {
@@ -670,8 +683,8 @@ ${conversationText}
       logger.debug(`[MemorySystem] 归档时间范围: ${new Date(lastMsg.timestamp).toISOString()} ~ ${new Date(firstMsg.timestamp).toISOString()}`);
       logger.debug(`[MemorySystem] 最旧消息ID: ${lastMsg.id}, 最新归档消息ID: ${firstMsg.id}`);
 
-      // 6. 匹配或创建 Topic
-      const existingTopics = await this.getTopics(userId, 5);
+      // 6. 匹配或创建 Topic（只匹配 active 状态的 Topic）
+      const existingTopics = await this.getTopics(userId, 5, 'active');
       const topicResult = await this.matchOrCreateTopic(summary, existingTopics);
 
       // 7. 执行归档
